# api/openapi.v1.yaml
...
components:
  schemas:
...
    ErrorBody:
      additionalProperties: false
      properties:
        code:
          title: Code
          type: string
        message:
          title: Message
          type: string
      required:
      - code
      - message
      title: ErrorBody
      type: object
    ErrorResponse:
      additionalProperties: false
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
      required:
      - error
      title: ErrorResponse
      type: object
...
    GuestExportResponse:
      additionalProperties: false
      description: Everything stored for one guest.
      properties:
        check_ins:
          items:
            additionalProperties: true
            type: object
          title: Check Ins
          type: array
        experiment_assignments:
          items:
            additionalProperties: true
            type: object
          title: Experiment Assignments
          type: array
        experiment_exposures:
          items:
            additionalProperties: true
            type: object
          title: Experiment Exposures
          type: array
        feedback:
          items:
            additionalProperties: true
            type: object
          title: Feedback
          type: array
        guest_id:
          format: uuid
          title: Guest Id
          type: string
        playback_events:
          items:
            additionalProperties: true
            type: object
          title: Playback Events
          type: array
        recommendations:
          items:
            additionalProperties: true
            type: object
          title: Recommendations
          type: array
        sessions:
          items:
            additionalProperties: true
            type: object
          title: Sessions
          type: array
      required:
      - guest_id
      - check_ins
      - recommendations
      - sessions
      - feedback
      - experiment_assignments
      - experiment_exposures
      - playback_events
      title: GuestExportResponse
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    SessionHistoryPage:
      additionalProperties: false
      properties:
        has_more:
          title: Has More
          type: boolean
        items:
          items:
            $ref: '#/components/schemas/SessionSummary'
          title: Items
          type: array
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
      required:
      - items
      - next_cursor
      - has_more
      title: SessionHistoryPage
      type: object
...
    SessionSummary:
      additionalProperties: false
      description: One row of a guest's history.
      properties:
        completed_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Completed At
        created_at:
          format: date-time
          title: Created At
          type: string
        duration_minutes:
          title: Duration Minutes
          type: integer
        id:
          format: uuid
          title: Id
          type: string
        outcome_score:
          anyOf:
          - type: integer
          - type: 'null'
          title: Outcome Score
        practice_id:
          title: Practice Id
          type: string
        public_title:
          title: Public Title
          type: string
        rule_set_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Rule Set Version
        status:
          title: Status
          type: string
      required:
      - id
      - status
      - created_at
      - completed_at
      - practice_id
      - public_title
      - duration_minutes
      - rule_set_version
      - outcome_score
      title: SessionSummary
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  description: 'Adaptive Meditation Program001 vertical slice: check-in, deterministic practice recommendation,
    session protocol and outcome feedback. Wellness product; it does not diagnose or treat any condition.'
  title: Adaptive Meditation API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /v1/me/data:
    delete:
      description: 'Deletes the guest row; the database cascades to everything else.


        One statement rather than a list of tables, so a table added later cannot be

        forgotten here.'
      operationId: deleteGuestData
      parameters:
      - in: header
        name: X-Guest-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Guest-Id
      responses:
        '204':
          description: Successful Response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Guest identity required
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unknown guest
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete everything stored for this guest
      tags:
      - guest
  /v1/me/export:
    get:
      operationId: exportGuestData
      parameters:
      - in: header
        name: X-Guest-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Guest-Id
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuestExportResponse'
          description: Successful Response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Guest identity required
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unknown guest
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Everything stored for this guest
      tags:
      - guest
...
  /v1/sessions/history:
    get:
      operationId: listSessionHistory
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          title: Limit
          type: integer
      - in: query
        name: cursor
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      - in: header
        name: X-Guest-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Guest-Id
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionHistoryPage'
          description: Successful Response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Guest identity required
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Validation error
      summary: This guest's past sessions, newest first
      tags:
      - guest
...